Remove (u)int64 math
We carefully calculated that there can't be any overflow.
There is also no loss in precision, it only needed
simple loss-less math simplifactions.

+++ software/src/adc.h
@@ -54,9 +54,9 @@ typedef struct {
 	XMC_VADC_GROUP_t *group;
 
 	// ADC result
-	uint64_t result_sum[2];
+	uint32_t result_sum[2];
 	uint32_t result_count[2];
-	int64_t result[2];
+	int32_t result[2];
 	int32_t result_mv[2];
 	uint8_t result_index[2];
